requirements.txt
uvloop==0.12
sanic==19.6.3
sanic-cors==0.9.9.post3
sanic-jwt-extended==0.3.1
gino==0.8.2
pytest-sanic==1.0.0
psycopg2-binary>=2.7.7
pandas==0.24.2
Click==7.0
alembic==1.0.9
pydantic==0.31
gunicorn==19.9.0
uvicorn==0.8.6
pillow==6.2.1